fix: ts 에러 수정
