+++ front/src/App.tsx
@@ -3,8 +3,8 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Layout from "./components/Layout";
 import Home from "./pages/Home";
 import AllContents from "./pages/AllContents";
-import Content from "./pages/Content";
-import NewContent from "./pages/NewContent";
+import Content from "./pages/Content/Content";
+import NewContent from "./pages/NewContent/NewContent";
 // import EditContent from "./pages/EditContent";
 import Register from "./pages/Register";
 import Login from "./pages/Login";
@@ -22,10 +22,10 @@ function App() {
             <Route path="new" element={<NewContent />} />
             <Route path="content/:id" element={<Content />} />
             {/* <Route path="edit" element={<EditContent />} /> */}
-            <Route path="register" element={<Register />} />
+            {/* <Route path="register" element={<Register />} /> */}
             <Route path="login" element={<Login />} />
             <Route path="mypage" element={<MyPage isMain={true} />} />
-            <Route path="mypage/edit" element={<MyPage />} />
+            {/* <Route path="mypage/edit" element={<MyPage />} /> */}
           </Route>
         </Routes>
       </div>
