+++ front/src/pages/EditContent.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const EditContent = () => {
+  return <div></div>;
+};
+
+export default EditContent;
