+++ front/src/pages/NewContent/NewContent.tsx
@@ -2,9 +2,16 @@ import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 // import Input from '../components/Input.jsx'
-import MyButton from "../components/MyButton";
-import { Title } from "./Register";
+import MyButton from "../../components/MyButton";
+import { Title } from "../Register";
 import axios from "axios";
+import {
+  TitleInput,
+  TitleWrap,
+  ContentInput,
+  ImgSearchInput,
+  ButtonWrap,
+} from "./NewContent.styles";
 
 function NewContent() {
   const navigate = useNavigate();
@@ -21,7 +28,7 @@ function NewContent() {
     headers: { Authorization: `Bearer ${token}` },
   };
 
-  const handleChangeState = (e) => {
+  const handleChangeState = (e: React.ChangeEvent<HTMLInputElement>) => {
     setState({ ...state, [e.target.name]: e.target.value });
   };
 
@@ -84,14 +91,14 @@ function NewContent() {
           value={state.img}
           onChange={handleChangeState}
         />
-        <MyButton text="검색" type="basic" />
+        {/* <MyButton text="검색" type="basic" /> */}
       </div>
       <p>내용</p>
       <ContentInput
         name="content"
         placeholder="내용을 적어주세요."
         value={state.content}
-        onChange={handleChangeState}
+        // onChange={handleChangeState}
       />
       <ButtonWrap>
         <MyButton text="저장하기" type="submit" onClick={handleSubmit} />
@@ -100,45 +107,5 @@ function NewContent() {
     </div>
   );
 }
-
-const TitleWrap = styled.div`
-  display: flex;
-  justify-content: center;
-`;
-
-const TitleInput = styled.input`
-  width: 500px;
-  height: 30px;
-  margin: 10px 0;
-  padding: 5px 10px;
-  border: 1px solid grey;
-  border-radius: 8px;
-  font-family: "S-CoreDream-4Regular";
-  &:focus {
-    outline: 2px solid #003c8f;
-  }
-`;
-
-const ContentInput = styled.textarea`
-  width: 500px;
-  height: 300px;
-  margin: 10px 0;
-  padding: 10px;
-  border: 1px solid grey;
-  border-radius: 8px;
-  font-family: "S-CoreDream-4Regular";
-  &:focus {
-    outline: 2px solid #003c8f;
-  }
-`;
-const ImgSearchInput = styled(TitleInput)`
-  width: 400px;
-  margin-right: 20px;
-`;
-
-export const ButtonWrap = styled.div`
-  display: flex;
-  justify-content: flex-end;
-`;
 
 export default NewContent;
